# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tenant Payment & Communication API",
    "description": "REST API for property management, lease handling, payment processing, and tenant communication. Includes user authentication, property management, payment tracking, maintenance requests, messaging, and AI-powered features.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/properties": {
      "get": {
        "tags": ["Properties"],
        "summary": "List properties",
        "description": "Get all properties for the authenticated landlord",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 10 },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Properties retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "properties": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Property" }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "total": { "type": "integer" },
                            "page": { "type": "integer" },
                            "limit": { "type": "integer" },
                            "pages": { "type": "integer" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Properties"],
        "summary": "Create property",
        "description": "Create a new property (landlord only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "address", "city", "state", "zipCode", "propertyType"],
                "properties": {
                  "name": { "type": "string", "example": "Sunset Apartments" },
                  "address": { "type": "string", "example": "123 Main Street" },
                  "city": { "type": "string", "example": "Los Angeles" },
                  "state": { "type": "string", "example": "CA" },
                  "zipCode": { "type": "string", "example": "90001" },
                  "propertyType": { "type": "string", "enum": ["apartment", "house", "condo", "townhouse", "commercial"] },
                  "unitCount": { "type": "integer", "example": 2 },
                  "description": { "type": "string" },
                  "amenities": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Property created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Property created successfully" },
                    "data": { "$ref": "#/components/schemas/Property" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token received from login endpoint"
      }
    },
    "schemas": {
...
      "Property": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "landlordId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "address": { "type": "string" },
          "city": { "type": "string" },
          "state": { "type": "string" },
          "zipCode": { "type": "string" },
          "propertyType": { "type": "string", "enum": ["apartment", "house", "condo", "townhouse", "commercial"] },
          "unitCount": { "type": "integer" },
          "description": { "type": "string" },
          "amenities": { "type": "array", "items": { "type": "string" } },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "code": { "type": "string" }
        }
      }
    }
  }
}